+++ package.json
@@ -18,21 +18,10 @@
     "start": "node server",
     "mocha": "node node_modules/.bin/mocha packages/**/server/tests/**/*.js -R spec -r tools/test/mocha-req.js",
     "karma": "node node_modules/karma/bin/karma start",
-    "test": "grunt test",
+    "test": "gulp test",
     "postinstall": "node node_modules/meanio/node_modules/mean-cli/bin/mean-postinstall"
   },
   "dependencies": {
-    "grunt-cli": "^0.1.13",
-    "grunt-hook": "~0.3.0",
-    "grunt-concurrent": "^1.0.0",
-    "grunt-contrib-clean": "^0.6.0",
-    "grunt-contrib-csslint": "^0.3.1",
-    "grunt-contrib-cssmin": "^0.10.0",
-    "grunt-contrib-jshint": "^0.10.0",
-    "grunt-contrib-uglify": "^0.5.1",
-    "grunt-contrib-watch": "^0.6.1",
-    "grunt-env": "^0.4.1",
-    "grunt-nodemon": "^0.3.0",
     "time-grunt": "^1.0.0",
     "assetmanager": "1.1.2",
     "async": "0.9.0",
